+++ CollisioniLib.vcxproj
@@ -21,16 +21,16 @@
   <ItemGroup>
     <ClInclude Include="include\Collision\Collider.h" />
     <ClInclude Include="include\Collision\CollisionManager.h" />
-    <ClInclude Include="include\sys\Mathematics.h" />
-    <ClInclude Include="include\sys\Singleton.h" />
-    <ClInclude Include="include\sys\System.h" />
+    <ClInclude Include="src\sys\Mathematics.h" />
+    <ClInclude Include="src\sys\Singleton.h" />
+    <ClInclude Include="src\sys\System.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="include\Collision\Collider.cpp" />
-    <ClCompile Include="include\Collision\CollisionManager.cpp" />
-    <ClCompile Include="include\sys\Mathematics.cpp" />
-    <ClCompile Include="include\sys\Singleton.cpp" />
-    <ClCompile Include="include\sys\System.cpp" />
+    <ClCompile Include="src\Collision\Collider.cpp" />
+    <ClCompile Include="src\Collision\CollisionManager.cpp" />
+    <ClCompile Include="src\sys\Mathematics.cpp" />
+    <ClCompile Include="src\sys\Singleton.cpp" />
+    <ClCompile Include="src\sys\System.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>17.0</VCProjectVersion>
@@ -131,7 +131,7 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_LIB;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>$(ProjectDir)include;$(SolutionDir)GameProject;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);$(ProjectDir)include;$(SolutionDir)GameProject;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <AdditionalOptions>/utf-8 %(AdditionalOptions)</AdditionalOptions>
       <RuntimeLibrary>MultiThreadedDebug</RuntimeLibrary>
@@ -150,7 +150,7 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>NDEBUG;_LIB;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>$(ProjectDir)include;$(SolutionDir)GameProject;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);$(ProjectDir)include;$(SolutionDir)GameProject;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <AdditionalOptions>/utf-8 %(AdditionalOptions)</AdditionalOptions>
       <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
